Injury Claim AI/requirements.txt
fastapi
uvicorn
python-docx
sqlalchemy
langchain
langchain-google-genai
google-generativeai
python-dotenv
tavily-python
